requirements.txt
djangorestframework>=2.3.14
# Additional requirements go here
